backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
supabase
qdrant-client
slowapi
python-dotenv
httpx
supabase_auth
